erratum: avoid another syncBugs() call
Similar to the change in df6e8be77212179a5a97ee350bb2d7c7821febae, we
have to remove this call because it's not implemented in our public
codebase right now.

As described in the other syncBugs() removal change, removing this here
might cause issues for RHOS's use of this class :( Need to think about
that further.